Use permutations to compare diachronic variations

More details here.

Usage

diachrony_uniform(
  df,
  tpq,
  taq,
  y,
  group,
  k = 10,
  predictor_fun = predictor_loess,
  x_prediction = 30,
  ...
)

Arguments

df

data.frame() to work on

tpq, taq

colname to use for tpq and taq (when shaking_gaussian)

y

colname to use for the value of interest

group

colname (optionnal) whether you have groups

k

number of permutations (10 by default)

predictor_fun

function one of predictor

x_prediction

on which to predict new values. See Details.

...

additional arguments to predictor_fun

Details

If x_prediction can be passed as a single numeric or as a vector of numeric. If a single numeric is passed, a regular sequence of this length will be created between min(tpq) and max(taq), ie covering the full temporal span. If a vector of numeric is passed, then these values will be used. By default, it eis 30, so that 30 dates covering the temporal span will be used. See examples.

Examples

set.seed(2329) # for replicability

# let's say animals had different names
x <- animals %>% dplyr::rename(post=tpq, ante=taq, lsi=value)
x
#> # A tibble: 5,533 × 6
#>    taxa  site   us        post  ante      lsi
#>    <fct> <chr>  <chr>    <dbl> <dbl>    <dbl>
#>  1 frog  Alz-Br A-B_FO    -125   -50 -0.0352 
#>  2 frog  Ans-lM A-M_12      50   700 -0.123  
#>  3 frog  Ans-lM A-M_12      50   700 -0.129  
#>  4 frog  Ans-lM A-M_139    150   200  0.120  
#>  5 frog  Ans-lM An-M_15     50   700  0.0575 
#>  6 frog  Ans-lM A-M_150     50   200  0.0528 
#>  7 frog  Ans-lM A-M_150     50   200 -0.00804
#>  8 frog  Arl-Cp A-C_1061   -75   -40 -0.0760 
#>  9 frog  Arl-Cp A-C_4043   -40   -15 -0.00385
#> 10 frog  Arl-Cp A-C_7039   -15    60 -0.0608 
#> # … with 5,523 more rows
x %>% diachrony_uniform(post, ante, lsi)
#>  * working on a temporal sequence of 30 steps
#>  * launching 10 permutations
#> # A tibble: 300 × 3
#>        k   x_new        y
#>    <int>   <dbl>    <dbl>
#>  1     1 -225    -0.132  
#>  2     1 -193.   -0.112  
#>  3     1 -161.   -0.0931 
#>  4     1 -129.   -0.0751 
#>  5     1  -97.4  -0.0582 
#>  6     1  -65.5  -0.0424 
#>  7     1  -33.6  -0.0280 
#>  8     1   -1.72 -0.0146 
#>  9     1   30.2  -0.00189
#> 10     1   62.1   0.00789
#> # … with 290 more rows

animals %>% diachrony_uniform(tpq, taq, value, k=2, group=taxa)
#>  * working on a temporal sequence of 30 steps
#>  * launching 2 permutations
#> # A tibble: 240 × 4
#>        k group   x_new        y
#>    <int> <chr>   <dbl>    <dbl>
#>  1     1 cat   -225    -0.165  
#>  2     1 cat   -193.   -0.155  
#>  3     1 cat   -161.   -0.142  
#>  4     1 cat   -129.   -0.126  
#>  5     1 cat    -97.4  -0.106  
#>  6     1 cat    -65.5  -0.0837 
#>  7     1 cat    -33.6  -0.0587 
#>  8     1 cat     -1.72 -0.0323 
#>  9     1 cat     30.2  -0.00544
#> 10     1 cat     62.1   0.0225 
#> # … with 230 more rows

if (FALSE) {
library(ggplot2)
x %>% ggplot() +
  aes(x_new, y, group=k, col=group) +
  geom_point(size=0.3) +
  theme_minimal() +
 geom_smooth(aes(group=NULL, col=group))
}